requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Django>=4.0,<5.0
psycopg2-binary>=2.9
python-decouple>=3.8
drf-yasg==1.21.8
djangorestframework==3.15.2
